Add Roles

+++ app/controllers/companies_controller.rb
@@ -2,22 +2,37 @@ class CompaniesController < ApplicationController
   before_action :authenticate_user!
 
   def index
-    @companies = Company.all
+    if current_user.role == 'owner'
+      @companies = Company.all.order(:name)
+    else
+      flash[:alert] = "You don't have access to this page!"
+      redirect_to root_path
+    end
   end
 
   def show
-    @company = Company.find(params[:id])
+    if CompanyUser.exists?(company_id: params[:id], user_id: current_user) || current_user.role == 'owner'
+      @company = Company.find(params[:id])
+    else
+      flash[:alert] = "You don't have access to this page!"
+      redirect_to root_path
+    end
   end
 
   def new
-    @company = Company.new
+    if !CompanyUser.exists?(user_id: current_user) && current_user.role == 'admin'
+      @company = Company.new
+    else
+      flash[:alert] = "You don't have access to this page!"
+      redirect_to root_path
+    end
   end
 
   def create
     @company = Company.new(company_params)
     @company.user = current_user
-    if @company.save
-      flash[:notice] = 'Company added successfully.'
+    if @company.save && CompanyUser.create(company_id: @company.id, user_id: current_user.id)
+      flash[:success] = 'Company added successfully.'
       redirect_to company_path(@company)
     else
       flash[:alert] = @company.errors.full_messages.join(', ')
@@ -26,13 +41,18 @@ class CompaniesController < ApplicationController
   end
 
   def edit
-    @company = Company.find(params[:id])
+    if (CompanyUser.exists?(company_id: params[:id], user_id: current_user) && current_user.role == 'admin') || current_user.role == 'owner'
+      @company = Company.find(params[:id])
+    else
+      flash[:alert] = "You don't have access to this page!"
+      redirect_to root_path
+    end
   end
 
   def update
     @company = Company.find(params[:id])
     if @company.update_attributes(company_params)
-      flash[:notice] = 'Company edited successfully.'
+      flash[:success] = 'Company edited successfully.'
       redirect_to company_path(@company)
     else
       flash[:alert] = @company.errors.full_messages.join(', ')
@@ -41,10 +61,16 @@ class CompaniesController < ApplicationController
   end
 
   def destroy
-    company = Company.find(params[:id])
-    company.destroy
-    flash[:notice] = 'Company deleted.'
-    redirect_to root_path
+    if current_user.role == 'admin' || current_user.role == 'owner'
+      company = Company.find(params[:id])
+      company.destroy
+      flash[:success] = 'Company deleted.'
+      if current_user.role == 'owner'
+        redirect_to companies_path
+      else
+        redirect_to root_path
+      end
+    end
   end
 
   protected
